codegen: Add JS view source generation
